+++ src/components/MinimizeView/MinimizeView.tsx
@@ -39,9 +39,16 @@ interface Props {
  * Minimize View
  */
 export const MinimizeView: React.FC<Props> = ({
-  options: { variable: variableName, padding = 0, emptyValue = false, persistent = false, customValue = false } = {},
+  options: {
+    variable: variableName,
+    padding = 0,
+    emptyValue = false,
+    persistent = false,
+    customValue = false,
+    showLabel = false,
+    labelWidth,
+  } = {},
   eventBus,
-  width,
   panelEventBus,
 }) => {
   /**
@@ -65,13 +72,6 @@ export const MinimizeView: React.FC<Props> = ({
     );
   }
 
-  /**
-   * Label and Select Width
-   */
-  const labelWidth = 10;
-  const labelWidthPx = labelWidth * 8;
-  const maxWidth = width - labelWidthPx - padding * 2;
-
   return (
     <div
       className={css`
@@ -79,18 +79,22 @@ export const MinimizeView: React.FC<Props> = ({
       `}
       data-testid={TEST_IDS.minimizeView.root}
     >
-      <InlineField className={styles.field} grow={true} label={variable.label || variable.name} labelWidth={labelWidth}>
+      <InlineField
+        className={styles.field}
+        grow={true}
+        shrink={true}
+        label={showLabel && (variable.label || variable.name)}
+        labelWidth={labelWidth}
+      >
         <>
           {(variable.type === VariableType.QUERY || variable.type === VariableType.CUSTOM) && (
-            <div style={{ maxWidth }}>
-              <OptionsVariable
-                variable={variable}
-                emptyValue={emptyValue}
-                persistent={persistent}
-                customValue={customValue}
-                panelEventBus={panelEventBus}
-              />
-            </div>
+            <OptionsVariable
+              variable={variable}
+              emptyValue={emptyValue}
+              persistent={persistent}
+              customValue={customValue}
+              panelEventBus={panelEventBus}
+            />
           )}
           {variable.type === VariableType.TEXTBOX && <TextVariable variable={variable} panelEventBus={panelEventBus} />}
         </>