+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Features / Enhancements
 
 - Update selecting and deselecting multi-levels (#139)
+- Update Auto/Static Minimize label width (#144)
+- Add show and hide Minimize label (#144)
 
 ## 2.5.0 (2024-04-01)